client/requirements.txt
flask
grpcio
grpcio-tools
asyncio